refactor: split command files
