+++ src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub mod build;
+pub mod upload;
+pub mod simulator;
